refactor: gather the SwiftUI layer under SwiftSync/UI/
Move the reactive/UI helpers into a self-contained UI/ folder and split the
phantom-named ReactiveQuery.swift (held SyncQuery + SyncModel, no such type)
into type-named SyncQuery.swift / SyncModel.swift, matching the publisher pair.

UI/: SyncQueryPublisher, SyncModelPublisher, SyncQuery, SyncModel,
SyncedQuery (+SyncedModel/driver/phase/results), SyncedView. Pure move/rename;
same module, no API or import change. SwiftSync 181 green.
